Renamed files and created src/
