+++ example/bot.js
@@ -1,6 +1,5 @@
-const Discord = require('discord.mybot');
+const Discord = require('../');
 const client = new Discord.Client();
 
 // Token bot en: https://discordapp.com/developers/applications
-client.login('BOT_TOKEN')
-
+client.login('BOT_TOKEN');
