+++ package.json
@@ -1,23 +1,32 @@
 {
-  "name": "discord.mybot",
-  "version": "0.0.1",
-  "description": "discord.mybot es un módulo node.js que le permite interactuar con la API de Discord con mucha facilidad.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mybot-team/discord.mybot.git"
-  },
-  "keywords": ["discord", "discord.js", "discord api", "api"],
-  "author": "MyBOT Team",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mybot-team/discord.mybot/issues"
-  },
-  "homepage": "https://github.com/mybot-team/discord.mybot#readme",
-  "dependencies": {
-    "ws": "^7.2.3"
-  }
+    "name": "discord.mybot",
+    "version": "0.0.1",
+    "description": "Módulo de Node.js que permite interactuar con la API de Discord con mucha facilidad",
+    "main": "src/index.js",
+    "scripts": {
+        "eslint": "eslint src/",
+        "test": "node examples/bot.js"
+    },
+    "dependencies": {
+        "ws": "^7.2.3"
+    },
+    "devDependencies": {},
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/mybot-team/discord.mybot.git"
+    },
+    "bugs": {
+        "url": "https://github.com/mybot-team/discord.mybot/issues"
+    },
+    "homepage": "https://github.com/mybot-team/discord.mybot",
+    "keywords": [
+        "bots",
+        "discord",
+        "websocket",
+        "discord api",
+        "bots discord",
+        "discord.mybot"
+    ],
+    "author": "MyBOT Team",
+    "license": "MIT"
 }
